Refator to allow reading DAQMx data when kTocDAQmxRawData is not set
